+++ rolldown.config.ts
@@ -1,5 +1,10 @@
 import { defineConfig } from './src/rolldown-config';
 
 export default defineConfig.bemedev({
-  excludesTS: ['**/*.tsx', '**/fixtures/**/*.ts'],
+  excludesTS: [
+    '**/*.tsx',
+    '**/fixtures/**/*.ts',
+    '**/fixtures.ts',
+    '**/*.fixtures.ts',
+  ],
 });