+++ src/build-tests/constants.ts
@@ -1,6 +1,6 @@
 export const FILE_ORIGIN = 'file';
 export const devKey = 'devDependencies';
-export const this1 = 'this-gen-1';
+export const THIS1 = 'this-gen-1';
 export const EXPORT_KEYS = ['main', 'typings', 'module'] as const;
 export const EXPORT_KEY = 'exports';
 export const PATH_OUT_DIR = 'compilerOptions.outDir';
@@ -10,11 +10,11 @@ export const TS_PATH = './tsconfig.json';
 export const PACKAGE_PATH = './package.json';
 export const TARBALL_FOLDER = '.pack';
 export const TARBALL_EXTENSION = '.tgz';
-export const TEARDOWN_COMMAND = `pnpm remove ${this1}`;
+export const TEARDOWN_COMMAND = `pnpm remove ${THIS1}`;
 
-export const FIXTURES = {
-  true: 'true',
-  false: 'false',
-  recursive: 'TEST_RECURSIVE',
-  vitest: 'VITEST_VSCODE',
-} as const;
+// export const FIXTURES = {
+//   true: 'true',
+//   false: 'false',
+//   recursive: 'TEST_RECURSIVE',
+//   vitest: 'VITEST_VSCODE',
+// } as const;