+++ src/build-tests/index.ts
@@ -1,2 +1,5 @@
 export * from './addTarball';
 export * from './cleanup';
+export { THIS1 } from './constants';
+export * from './import';
+export * from './types';